xe: jit: introduce dsl::layout_t
