+++ src/gpu/intel/conv/jit/tiler.cpp
@@ -432,7 +432,7 @@ int inner_block(
     int dim_idx = tensor_conv_dim_index(dim, tensor_kind);
     gpu_assert(dim_idx != -1);
     auto &layout = compute_layout(cfg, tensor_kind);
-    return into<int>(inner_block(
+    return into<int>(jit::inner_block(
             layout, dim_idx, /*skip_outer=*/true, /*inner_only=*/false));
 }
