+++ src/gpu/intel/jit/dsl/tensor.hpp
@@ -490,6 +490,268 @@ inline icoord_t operator+(const icoord_t &a, const icoord_t &b) {
     return ret;
 }
 
+struct layout_t;
+
+namespace layout {
+using jit::operator<<;
+struct block_t {
+    block_t() = default;
+    block_t(const idx_t &dim, int64_t block, stride_t stride = stride_t())
+        : dim(dim), block(block), stride(stride) {}
+
+    bool operator==(const block_t &other) const {
+        return (dim == other.dim) && (block == other.block)
+                && (stride == other.stride);
+    }
+    bool operator!=(const block_t &other) const { return !(*this == other); }
+
+    size_t get_hash() const { return ir_utils::get_hash(dim, block, stride); }
+
+    std::string str() const {
+        std::ostringstream oss;
+        oss << "block_t(dim = " << dim;
+        oss << ", block = " << block;
+        oss << ", stride = " << stride.str();
+        oss << ")";
+        return oss.str();
+    }
+
+    idx_t dim;
+    int64_t block = 1; // Block size.
+    stride_t stride; // Stride between elements of the block.
+};
+
+std::vector<block_t> normalize_blocks(
+        const std::vector<block_t> &blocks, bool remove_size_1_blocks = true);
+
+// Iterates through subtiles of the layout by returning the coordinates for
+// each tile. The iteration order is defined by the layout blocks.
+struct tile_iterator_t {
+    tile_iterator_t &operator++();
+    const icoord_t &operator*() const { return coord_; }
+    tile_iterator_t begin() const { return *this; }
+    tile_iterator_t end() const { return {}; }
+    bool operator==(const tile_iterator_t &o) const { return d_ == o.d_; }
+    bool operator!=(const tile_iterator_t &o) const { return d_ != o.d_; }
+
+protected:
+    friend struct dnnl::impl::gpu::intel::jit::dsl::layout_t;
+    tile_iterator_t() = default;
+    tile_iterator_t(const layout_t &layout, const tile_t &tile);
+
+private:
+    struct index_data_t {
+        index_data_t(idx_t idx, int64_t block, int64_t stride_, int64_t tile)
+            : i(0)
+            , idx(idx)
+            , end(std::min(block, utils::div_up(stride_ * block, tile)))
+            , stride(stride_ * utils::div_up(block, end)) {}
+        bool operator==(const index_data_t &other) const {
+            return i == other.i && idx == other.idx && end == other.end
+                    && stride == other.stride;
+        }
+        int64_t i;
+        idx_t idx;
+        int64_t end;
+        int64_t stride;
+    };
+    std::vector<index_data_t> d_;
+    icoord_t coord_;
+};
+
+} // namespace layout
+
+struct layout_t {
+    using block_t = layout::block_t;
+
+    layout_t() : type_(type_t::undef()), ndims_(0), offset_(0) {
+        sanity_check();
+    }
+    layout_t(const type_t &type, const std::vector<int64_t> &dims,
+            const expr_t &offset = 0, bool do_normalize = true);
+    layout_t(const type_t &type, const std::vector<block_t> &blocks = {},
+            const expr_t &offset = 0, size_t ndims = max_ndims,
+            bool do_normalize = true);
+
+    layout_t with(const std::vector<block_t> &blocks,
+            bool do_normalize = true) const {
+        return {type_, blocks, offset_, ndims_, do_normalize};
+    }
+
+    layout_t with(const type_t &new_type) const {
+        return {new_type, blocks_, offset_, ndims_, false};
+    }
+
+    // Unknown/undefined strides are assumed to be the outermost block.
+    // Furthermore, undefined strides are assumed to be dense with respect to
+    // the previous block.
+    layout_t with_block(block_t block) const;
+
+    bool is_empty() const {
+        if (type_.is_undef()) gpu_assert(*this == layout_t());
+        return type_.is_undef();
+    }
+
+    // Use of this interface is deprecated.
+    size_t ndims(bool check_invalid = true) const {
+        if (check_invalid) gpu_assert(has_ndims());
+        return ndims_;
+    }
+
+    // Number of elements in the layout
+    int64_t elems(const idx_t &dim = {}) const {
+        int64_t ret = 1;
+        for (auto &b : blocks_)
+            if (dim.is_undef() || b.dim == dim) ret *= b.block;
+        return ret;
+    }
+
+    template <typename T = expr_t>
+    T offset(const coord_t &args = {}, bool ignore_offset = false) const;
+
+    const type_t &type() const { return type_; }
+
+    tile_t tile() const {
+        tile_t tile;
+        for (auto &b : blocks_)
+            tile[b.dim] = tile.get(b.dim, 1) * b.block;
+        return tile;
+    }
+
+    // Returns the maximum inner subtile containing less than max elements. This
+    // tile is dense within the layout when is_dense is true and perfectly
+    // subdivides the layout when perfectly_divides is true.
+    tile_t max_subtile(int64_t max, bool is_dense = true,
+            bool perfectly_divides = true) const;
+
+    stride_t stride(const idx_t &dim, int dim_block_idx = 0) const {
+        int idx = 0;
+        for (auto &b : blocks_) {
+            if (b.dim != dim) continue;
+            if (idx == dim_block_idx) { return b.stride; }
+            idx++;
+        }
+        return stride_t();
+    }
+
+    size_t nblocks() const { return blocks().size(); }
+
+    const std::vector<block_t> &blocks() const { return blocks_; }
+
+    const block_t &operator[](size_t idx) const { return blocks_[idx]; }
+    block_t &operator[](size_t idx) { return blocks_[idx]; }
+
+    void set_offset(const expr_t &offset) { offset_ = offset; }
+
+    bool is_strictly_equal(const layout_t &other, bool compare_offset = true,
+            bool compare_strides = true) const;
+
+    bool operator==(const layout_t &other) const {
+        return type_ == other.type_ && ndims_ == other.ndims_
+                && offset_.is_equal(other.offset_) && blocks_ == other.blocks_;
+    }
+    bool operator!=(const layout_t &other) const { return !operator==(other); }
+    bool operator<=(const layout_t &other) const;
+
+    bool is_equal_normalized(
+            const layout_t &other, bool compare_offset = true) const {
+        return normalize().is_strictly_equal(other.normalize(), compare_offset);
+    }
+
+    size_t get_hash() const {
+        return ir_utils::get_hash(type_, ndims_, offset_, blocks_);
+    }
+
+    expr_t operator()(const coord_t &coord) const { return offset(coord); }
+
+    std::string str() const {
+        if (is_const(offset(), 0)) return desc_str();
+        return desc_str() + " offset: " + offset_.str();
+    }
+
+    void dump() const { printf("%s\n", str().c_str()); }
+
+    // Returns a canonical representation of the layout:
+    // - Size one blocks are removed
+    // - Consecutive dense blocks are merged
+    layout_t normalize() const { return with(blocks_); }
+
+    // Returns a new (sub-)layout that fully contains the passed sub-tensor.
+    // Strides are kept unchanged.
+    // Assumption: the original layout can be tiled by the passed sub-tensor.
+    // For example: XaYb4a2b can be tiled into 2x2 sub-tensors but it's not
+    // possible to tile it into 3x2 sub-tensors.
+    layout_t sub(const tile_t &tile, const coord_t &start = {}) const;
+
+    bool is_dense() const {
+        stride_t stride = 1;
+        for (auto &b : blocks_) {
+            if (b.stride != stride) return false;
+            stride *= b.block;
+        }
+        return true;
+    }
+
+    // Returns a packed layout where all blocks are contiguous, without gaps.
+    layout_t make_dense() const {
+        int64_t stride = 1;
+        auto new_blocks = blocks_;
+        for (auto &b : new_blocks) {
+            b.stride = stride;
+            stride *= b.block;
+        }
+        return with(new_blocks);
+    }
+
+    // Returns an equivalent layout where the specified block is split into two.
+    // block0 - inner block size.
+    // block1 - outer block size.
+    layout_t split_block(
+            const block_t &b, int64_t block0, int64_t block1) const;
+
+    using tile_iterator_t = layout::tile_iterator_t;
+    tile_iterator_t iter(const tile_t &tile) const {
+        return tile_iterator_t(*this, tile);
+    };
+
+    template <typename F>
+    void for_each_tile(const tile_t &tile, const F &f) const {
+        for (auto &coord : iter(tile)) {
+            f(coord);
+        }
+    }
+
+    size_t get_idx(const block_t &b) const {
+        gpu_assert(&blocks().front() <= &b && &b <= &blocks().back());
+        return &b - &blocks().front();
+    }
+
+    bool is_outermost(const block_t &block) const {
+        for (size_t i = get_idx(block) + 1; i < blocks().size(); i++) {
+            if (blocks()[i].dim == block.dim) return false;
+        }
+        return true;
+    }
+
+private:
+    static constexpr size_t max_ndims = 16;
+    bool has_ndims() const { return ndims_ < max_ndims; }
+    std::string desc_str(bool dnnl_style = false) const;
+    void sanity_check() const;
+
+    type_t type_; // Data type of the layout.
+    size_t ndims_ = max_ndims; //(Deprecated) Number of dimensions.
+    expr_t offset_; // Offset to the start of the layout.
+    std::vector<block_t> blocks_; // Blocks ordered from innermost to outermost.
+};
+
+extern template expr_t layout_t::offset<expr_t>(
+        const coord_t &args, bool ignore_offset) const;
+extern template int layout_t::offset<int>(
+        const coord_t &args, bool ignore_offset) const;
+extern template int64_t layout_t::offset<int64_t>(
+        const coord_t &args, bool ignore_offset) const;
+
 } // namespace dsl
 } // namespace jit
 } // namespace intel
