+++ src/gpu/intel/jit/ir/object.cpp
@@ -89,6 +89,10 @@ expr_t::expr_t(uint64_t value) : object_t(new int_imm_t(value)) {}
 bool is_const(const expr_t &e) {
     return e.is<bool_imm_t>() || e.is<int_imm_t>() || e.is<float_imm_t>();
 }
+bool is_const(const expr_t &e, int value) {
+    if (!is_const(e)) return false;
+    return e.is_equal(to_expr(value, e.type()));
+}
 
 bool to_bool(const expr_t &e) {
     return to_cpp<bool>(e);
