+++ src/gpu/intel/jit/ir/tensor.hpp
@@ -214,441 +214,8 @@ private:
     dim_t cur_stride_;
 };
 
-struct layout_block_t {
-    layout_block_t() = default;
-
-    layout_block_t(const pvar_t &dim, int64_t block) : dim(dim), block(block) {}
-
-    layout_block_t(const pvar_t &dim, int64_t block, const stride_t &stride)
-        : dim(dim), block(block), stride(stride) {}
-
-    bool operator==(const layout_block_t &other) const {
-        return (dim == other.dim) && (block == other.block)
-                && (stride == other.stride);
-    }
-    bool operator!=(const layout_block_t &other) const {
-        return !(*this == other);
-    }
-
-    size_t get_hash() const { return ir_utils::get_hash(dim, block, stride); }
-
-    std::string str() const {
-        std::ostringstream oss;
-        oss << "block_t(dim = " << dim;
-        oss << ", block = " << block;
-        oss << ", stride = " << stride.str();
-        oss << ")";
-        return oss.str();
-    }
-
-    pvar_t dim;
-    int64_t block = 1; // Block size.
-    stride_t stride; // Stride between elements of the block.
-};
-
-std::vector<layout_block_t> normalize_blocks(
-        const std::vector<layout_block_t> &blocks,
-        bool remove_size_1_blocks = true);
-
-class layout_t {
-public:
-    layout_t() : type_(type_t::undef()), ndims_(0), offset_(0) {
-        sanity_check();
-    }
-
-    layout_t(const type_t &type, const std::vector<int64_t> &dims,
-            const expr_t &offset = 0, bool do_normalize = true)
-        : type_(type), ndims_(dims.size()), offset_(offset) {
-        if (type.is_undef()) {
-            *this = layout_t();
-            return;
-        }
-
-        int64_t stride = 1;
-        for (int64_t i = ndims_ - 1; i >= 0; i--) {
-            blocks_.emplace_back(i, dims[i], stride);
-            stride *= dims[i];
-        }
-        if (do_normalize) blocks_ = normalize_blocks(blocks_);
-        sanity_check();
-    }
-
-    layout_t(const type_t &type, const std::vector<layout_block_t> &blocks = {},
-            const expr_t &offset = 0, size_t ndims = max_ndims,
-            bool do_normalize = true)
-        : type_(type), ndims_(ndims), offset_(offset), blocks_(blocks) {
-        if (type.is_undef()) {
-            *this = layout_t();
-            return;
-        }
-
-        stride_t stride(1);
-        for (auto &b : blocks_) {
-            if (b.stride.is_undefined()) {
-                b.stride = stride;
-            } else {
-                stride = b.block;
-            }
-            stride *= b.block;
-        }
-        if (do_normalize) blocks_ = normalize_blocks(blocks_);
-        sanity_check();
-    }
-
-    layout_t with(std::vector<layout_block_t> blocks,
-            bool do_normalize = true) const {
-        layout_t ret = *this;
-        ret.blocks_
-                = do_normalize ? normalize_blocks(blocks) : std::move(blocks);
-        return ret;
-    }
-
-    layout_t with(const type_t &new_type) const {
-        auto ret = *this;
-        ret.type_ = new_type;
-        return ret;
-    }
-
-    // Unknown/undefined strides are assumed to be the outermost block.
-    // Furthermore, undefined strides are assumed to be dense with respect to
-    // the previous block.
-    layout_t with_block(layout_block_t block) const {
-        auto new_blocks = blocks();
-        if (block.stride.is_unknown()) {
-            new_blocks.emplace_back(block);
-        } else if (block.stride.is_undefined()) {
-            block.stride = !new_blocks.empty()
-                    ? new_blocks.back().stride * new_blocks.back().block
-                    : stride_t(1);
-            new_blocks.emplace_back(block);
-        } else {
-            auto it = new_blocks.begin();
-            while (it != new_blocks.end() && it->stride <= block.stride) {
-                it++;
-            }
-            new_blocks.insert(it, block);
-        }
-
-        auto ret = with(new_blocks);
-        if (ret.has_ndims()) {
-            if (block.dim.index() == ret.ndims()) ret.ndims_++;
-            gpu_assert(has_ndims());
-        }
-        return ret;
-    }
-
-    bool is_empty() const {
-        if (type_.is_undef()) gpu_assert(*this == layout_t());
-        return type_.is_undef();
-    }
-
-    size_t ndims(bool check_invalid = true) const {
-        if (check_invalid) gpu_assert(has_ndims());
-        return ndims_;
-    }
-
-    // Number of elements in the layout
-    int64_t elems(const pvar_t &dim = {}) const {
-        int64_t ret = 1;
-        for (auto &b : blocks_) {
-            if (!dim.is_undef() && b.dim != dim) continue;
-            ret *= b.block;
-        }
-        return ret;
-    }
-
-    template <typename T = expr_t>
-    T offset(const coord_t &args = {}, bool ignore_offset = false) const {
-        if (args.is_empty()) return expr_cast<T>(offset_);
-
-        expr_t off = 0;
-        auto _args = args;
-        for (auto &b : blocks()) {
-            if (!_args.has(b.dim)) continue;
-            auto &idx = _args[b.dim];
-            if (is_zero(idx)) continue;
-
-            // Do not use modulus for outermost blocks.
-            auto i = is_outermost(b) ? idx : (idx % b.block);
-            off = i * int64_t(b.stride) + off;
-            idx /= b.block;
-        }
-        if (ignore_offset) return expr_cast<T>(off);
-        return expr_cast<T>(offset_ + off);
-    }
-
-    const type_t &type() const { return type_; }
-
-    tile_t tile() const {
-        tile_t tile;
-        for (auto &b : blocks_)
-            tile[b.dim] = tile.get(b.dim, 1) * b.block;
-        return tile;
-    }
-
-    // Returns the maximum inner subtile containing less than max elements. This
-    // tile is dense within the layout when is_dense is true and perfectly
-    // subdivides the layout when perfectly_divides is true.
-    tile_t max_subtile(int64_t max, bool is_dense = true,
-            bool perfectly_divides = true) const;
-
-    stride_t stride(const pvar_t &dim, int dim_block_idx = 0) const {
-        int idx = 0;
-        for (auto &b : blocks_) {
-            if (b.dim != dim) continue;
-            if (idx == dim_block_idx) { return b.stride; }
-            idx++;
-        }
-        return stride_t();
-    }
-
-    size_t nblocks() const { return blocks().size(); }
-
-    const std::vector<layout_block_t> &blocks() const { return blocks_; }
-    std::vector<layout_block_t> &blocks() { return blocks_; }
-
-    const layout_block_t &operator[](size_t idx) const { return blocks_[idx]; }
-    layout_block_t &operator[](size_t idx) { return blocks_[idx]; }
-
-    void set_offset(const expr_t &offset) { offset_ = offset; }
-
-    bool is_strictly_equal(const layout_t &other, bool compare_offset = true,
-            bool compare_strides = true) const {
-        if (type_ != other.type_) return false;
-        if (compare_offset && !offset_.is_equal(other.offset_)) return false;
-        if (blocks_.size() != other.blocks_.size()) return false;
-        for (size_t i = 0; i < blocks_.size(); i++) {
-            auto &b0 = blocks_[i];
-            auto &b1 = other.blocks_[i];
-            if (b0.dim != b1.dim) return false;
-            if (b0.block != b1.block) return false;
-            if (compare_strides && b0.stride != b1.stride) return false;
-        }
-        return true;
-    }
-
-    bool operator==(const layout_t &other) const {
-        return type_ == other.type_ && ndims_ == other.ndims_
-                && offset_.is_equal(other.offset_) && blocks_ == other.blocks_;
-    }
-    bool operator!=(const layout_t &other) const { return !operator==(other); }
-    bool operator<=(const layout_t &other) const {
-        if (type_ != other.type_) return false;
-        auto other_blocks = other.normalize().blocks();
-        auto self_blocks = normalize().blocks();
-        if (self_blocks.size() > other_blocks.size()) return false;
-        if (self_blocks.empty()) return true;
-
-        int i = 0;
-        for (; i < (int)self_blocks.size() - 1; i++) {
-            if (self_blocks[i] != other_blocks[i]) return false;
-        }
-        return (self_blocks[i].dim == other_blocks[i].dim
-                && self_blocks[i].stride == other_blocks[i].stride
-                && other_blocks[i].block % self_blocks[i].block == 0);
-    }
-
-    bool is_equal_normalized(
-            const layout_t &other, bool compare_offset = true) const {
-        return normalize().is_strictly_equal(other.normalize(), compare_offset);
-    }
-
-    size_t get_hash() const {
-        return ir_utils::get_hash(type_, ndims_, offset_, blocks_);
-    }
-
-    expr_t operator()(const coord_t &coord) const { return offset(coord); }
-
-    std::string desc_str(bool dnnl_style = false) const {
-        if (is_empty()) return "(nil)";
-        if (!dnnl_style && blocks_.empty())
-            return "(scalar:" + type().str() + ")";
-
-        auto to_str = [](const pvar_t &dim, bool is_outer) {
-            auto ret = dim.str();
-            if (ret.length() == 1) {
-                if (is_outer) ret[0] -= 'a' - 'A';
-                return ret;
-            }
-            return "<" + ret + ">";
-        };
-        std::string ret;
-        stride_t dense_stride(1);
-        pvar_map_t<bool> seen;
-        for (auto &b : blocks()) {
-            std::string b_str;
-            if (dnnl_style && is_outermost(b)) {
-                b_str += to_str(b.dim, seen.get(b.dim, false));
-            } else {
-                b_str = std::to_string(b.block);
-                b_str += to_str(b.dim, false);
-            }
-            if (!dnnl_style) {
-                if (b.stride.is_unknown()) {
-                    b_str.append(1, '?');
-                } else if (b.stride != dense_stride) {
-                    b_str.append(1, '*');
-                }
-            }
-            b_str += ret;
-            std::swap(ret, b_str);
-            dense_stride = b.stride * b.block;
-            seen[b.dim] = true;
-        }
-        ret += ":" + type().str();
-        return ret;
-    }
-
-    std::string str() const {
-        if (is_empty()) return "(nil)";
-        ostringstream_t oss;
-        oss << desc_str();
-        if (!is_zero(offset())) oss << " offset: " << offset_;
-        return oss.str();
-    }
-
-    IR_DEFINE_DUMP()
-
-    // Returns a canonical representation of the layout:
-    // - Size one blocks are removed
-    // - Consecutive dense blocks are merged
-    layout_t normalize() const { return with(blocks_); }
-
-    // Returns a new (sub-)layout that fully contains the passed sub-tensor.
-    // Strides are kept unchanged.
-    // Assumption: the original layout can be tiled by the passed sub-tensor.
-    // For example: XaYb4a2b can be tiled into 2x2 sub-tensors but it's not
-    // possible to tile it into 3x2 sub-tensors.
-    layout_t sub(const tile_t &tile, const coord_t &start = {}) const;
-
-    bool is_dense() const {
-        stride_t stride = 1;
-        for (auto &b : blocks_) {
-            if (b.stride != stride) return false;
-            stride *= b.block;
-        }
-        return true;
-    }
-
-    // Returns a packed layout where all blocks are contiguous, without gaps.
-    layout_t make_dense() const {
-        int64_t stride = 1;
-        auto new_blocks = blocks_;
-        for (auto &b : new_blocks) {
-            b.stride = stride;
-            stride *= b.block;
-        }
-        return with(new_blocks);
-    }
-
-    // Returns an equivalent layout where the specified block is split into two.
-    // block0 - inner block size.
-    // block1 - outer block size.
-    layout_t split_block(
-            const layout_block_t &b, int64_t block0, int64_t block1) const;
-
-    // Iterates through subtiles of the layout by returning the coordinates for
-    // each tile. The iteration order is defined by the layout blocks.
-    struct tile_iterator_t {
-        tile_iterator_t &operator++() {
-            for (size_t i = 0; i < d_.size(); i++) {
-                if (d_[i].i < d_[i].end - 1) {
-                    coord_[d_[i].idx] += d_[i].stride;
-                    d_[i].i++;
-                    return *this;
-                }
-                coord_[d_[i].idx] -= d_[i].i * d_[i].stride;
-                d_[i].i = 0;
-            }
-            *this = end();
-            return *this;
-        }
-
-        const icoord_t &operator*() const { return coord_; }
-        tile_iterator_t begin() const { return *this; }
-        tile_iterator_t end() const { return {}; }
-        bool operator==(const tile_iterator_t &o) const { return d_ == o.d_; }
-        bool operator!=(const tile_iterator_t &o) const { return d_ != o.d_; }
-
-    protected:
-        friend class layout_t;
-        tile_iterator_t() = default;
-        tile_iterator_t(const layout_t &layout, const tile_t &tile) {
-            tile_t strides;
-            for (auto &b : layout.blocks()) {
-                auto &stride = strides[b.dim];
-                d_.emplace_back(b.dim, b.block, stride, tile.get(b.dim));
-                coord_[b.dim] = 0;
-                stride *= b.block;
-            }
-            if (!layout.is_empty() && layout.blocks().empty()) {
-                d_.emplace_back(*tile.begin(), 1, 1, 1);
-                coord_[d_[0].idx] = 0;
-            }
-        }
-
-    private:
-        struct index_data_t {
-            index_data_t(
-                    pvar_t idx, int64_t block, int64_t stride_, int64_t tile)
-                : i(0)
-                , idx(idx)
-                , end(std::min(block, utils::div_up(stride_ * block, tile)))
-                , stride(stride_ * utils::div_up(block, end)) {}
-            bool operator==(const index_data_t &other) const {
-                return i == other.i && idx == other.idx && end == other.end
-                        && stride == other.stride;
-            }
-            int64_t i;
-            pvar_t idx;
-            int64_t end;
-            int64_t stride;
-        };
-        std::vector<index_data_t> d_;
-        icoord_t coord_;
-    };
-
-    tile_iterator_t iter(const tile_t &tile) const {
-        return tile_iterator_t(*this, tile);
-    };
-
-    template <typename F>
-    void for_each_tile(const tile_t &tile, const F &f) const {
-        for (auto &coord : iter(tile)) {
-            f(coord);
-        }
-    }
-
-    size_t get_idx(const layout_block_t &b) const {
-        gpu_assert(&blocks().front() <= &b && &b <= &blocks().back());
-        return &b - &blocks().front();
-    }
-
-    bool is_outermost(const layout_block_t &block) const {
-        for (size_t i = get_idx(block) + 1; i < blocks().size(); i++) {
-            if (blocks()[i].dim == block.dim) return false;
-        }
-        return true;
-    }
-
-private:
-    static constexpr size_t max_ndims = 16;
-    bool has_ndims() const { return ndims_ < max_ndims; }
-
-    void sanity_check() const;
-
-    // Data type of the layout.
-    type_t type_;
-
-    // Number of dimensions.
-    size_t ndims_ = max_ndims;
-
-    // Offset to the start of the layout (in elements of type).
-    expr_t offset_;
-
-    // Blocks ordered from innermost to outermost.
-    std::vector<layout_block_t> blocks_;
-};
+using layout_block_t = dsl::layout::block_t;
+using layout_t = dsl::layout_t;
 
 inline dim_t inner_block(const layout_t &layout, const pvar_t &dim,
         bool skip_outer = true, bool inner_only = true) {
@@ -1301,7 +868,7 @@ public:
 
     layout_t normalized_tlayout() const {
         auto blocks = move_size_1_blocks_outer();
-        blocks = normalize_blocks(blocks, false);
+        blocks = dsl::layout::normalize_blocks(blocks, false);
         auto layout = tlayout_.with(blocks, false);
         return layout;
     }
